FLASH_ENG/backend/requirements.txt
pymongo==4.10.1
python-dotenv==1.0.1
PyJWT==2.9.0
bcrypt==4.2.1